feat: publications functionalities

+++ src/media/media.service.ts
@@ -1,12 +1,17 @@
-import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Inject, Injectable, forwardRef } from '@nestjs/common';
 import { CreateMediaDto } from './dto/create-media.dto';
 import { UpdateMediaDto } from './dto/update-media.dto';
 import { MediaRepository } from './media.reposirory';
+import { PublicationService } from 'src/publication/publication.service';
 
 
 @Injectable()
 export class MediaService {
-  constructor(private readonly mediaRepository: MediaRepository){}
+  constructor(
+    private readonly mediaRepository: MediaRepository,
+    @Inject(forwardRef(() =>  PublicationService))
+    private readonly publicationService: PublicationService
+    ){}
 
   async create(createMediaDto: CreateMediaDto) {
 
@@ -33,7 +38,7 @@ export class MediaService {
     if (!mediaAccount) {
       throw new HttpException('Media Account Not Found', HttpStatus.NOT_FOUND)
     }
-    return mediaAccount;// verificar se tem q retornar dentro de array
+    return mediaAccount;
   }
 
   async update(id: number, updateMediaDto: UpdateMediaDto) {
@@ -49,14 +54,20 @@ export class MediaService {
       throw new HttpException('These Credentials Are Already Taken', HttpStatus.CONFLICT)
     }
 
-    return this.mediaRepository.update(id, updateMediaDto);
+    return await this.mediaRepository.update(id, updateMediaDto);
   }
 
   async remove(id: number) {
-    const isMedia = this.mediaRepository.findOne(id);
+    const isMedia = await this.findOne(id);
     if (!isMedia) {
       throw new HttpException('No Register Found For The Given Id', HttpStatus.NOT_FOUND)
     }
+
+    const isPublication = await this.publicationService.findOneByMediaId(id);
+    if (isPublication) {
+      throw new HttpException(`This Media Cannot be Deleted. It was either scheduled or published`, HttpStatus.FORBIDDEN)
+    }
+
     return await this.mediaRepository.remove(id);
   }
